// SDKs/twilio.com/twilio_conversations_v1/1.36.0/typescript/src/sdk/models/shared/conversationsv1serviceservicerole.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { ServiceRoleEnumRoleTypeEnum } from "./serviceroleenumroletypeenum";


export class ConversationsV1ServiceServiceRole extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=chat_service_sid" })
  chatServiceSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=permissions" })
  permissions?: string[];

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=type" })
  type?: ServiceRoleEnumRoleTypeEnum;

  @Metadata({ data: "json, name=url" })
  url?: string;
}